add support for dynamic trace
Adds an optional TraceProvider to the Config which will be used to
extract a Trace from the current Context. This allows for using a
standard telemetry library like otel by providing a wrapper to translate
from their context to slogdriver's Trace. This has the benefit that
slogdriver does not need to be notified if new trace spans are added
through the telemetry library.